WD-26946 - Move tracing and metrics to opentelemetry module
